hackerearth indeed
You are given a string that contains only three characters , , and . You can change at the most  characters in the string. The uniformity index of a string is defined by the maximum length of the substring that contains the same character. Your task is to determine the maximum uniformity index than can be achieved